backend/requirements.txt
# Backend Dependencies for Candle-Light

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
asyncpg==0.30.0
alembic==1.14.0
greenlet==3.1.1

# Pydantic & Validation
pydantic==2.10.0
pydantic-settings==2.6.0
email-validator==2.2.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.2
httpx==0.27.2

# AI/ML
openai==1.50.0
google-generativeai==0.8.0
pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Production
gunicorn==23.0.0
